fix compilation

+++ OTRExporter/DisplayListExporter.cpp
@@ -251,7 +251,7 @@ void OTRExporter_DisplayList::Save(ZResource* res, fs::path outPath, BinaryWrite
 					spdlog::error(StringHelper::Sprintf("dListDecl == nullptr! Addr = %08X", GETSEGOFFSET(data)));
 				}
 
-				for (int i = 0; i < dList->otherDLists.size(); i++)
+				for (size_t i = 0; i < dList->otherDLists.size(); i++)
 				{
 					Declaration* dListDecl2 = dList->parent->GetDeclaration(GETSEGOFFSET(dList->otherDLists[i]->GetRawDataIndex()));
